### 2019-07-25 #1595 @kyleniemeyer
Hi @nuest, while we get started handling this submission, please merge the PR I submitted, which fixes the authors field in the paper: https://github.com/o2r-project/containerit/pull/152
