### 2019-07-25 #1595 @kyleniemeyer
@arfon whedon didn't find a license, but I can see the GPL-3.0 license in the repository: https://github.com/o2r-project/containerit/blob/master/LICENSE

Any ideas? Just ignore?
