### 2019-07-25 #1595 @kyleniemeyer
@lheagy @kthyng  can one of you edit this submission? I realize it's a bit outside your domain, but others are a bit overloaded and Karthik (who I'd normally ship R stuff to) is away for a bit.
