### 2019-07-26 #1595 @arfon
> @arfon whedon didn't find a license, but I can see the GPL-3.0 license in the repository: https://github.com/o2r-project/containerit/blob/master/LICENSE

> Any ideas? Just ignore?

Weird, I think this can be safely ignored.
