### 2019-07-26 #1595 @nuest
Hi! Thanks for getting things going, and for fixing the error in the `paper.md` rightaway.
