### 2019-07-26 16bc225
Add manual breaks to code snippet in paper.md
