### 2019-07-26 #1595 @nuest
PDF generation works, and I added some manual breaks in the contained code snippet so the line does not run off the page:

![image](https://user-images.githubusercontent.com/1325054/61932260-2d791f80-af83-11e9-9434-d8140831aab5.png)

**Now**:

![image](https://user-images.githubusercontent.com/1325054/61933814-0c1a3280-af87-11e9-91d0-55a608a3f5ed.png)
