### 2019-07-26 #1595 @nuest
**Suggested reviewers** from the JOSS list:

- @cole-brokamp
- @vsoch
- @jsta
- @adamhsparks
