### 2019-07-26 #1595 @cole-brokamp
I'm currently reviewing a singularity compose manuscript by @vsoch but I can't turn this one down either!  I would love to review as well if I can!
