### 2019-07-26 #1595 @kyleniemeyer
Thanks @vsoch and @cole-brokamp! We're still waiting to get an editor assigned, but I will definitely take you up on the offers to review. @cole-brokamp you can certainly wait until you have completed your other review.
