### 2019-07-27 #1595 @adamhsparks
I'm stupidly busy right now, already reviewing for rOpenSci and prepping
for some international travel in a few weeks. So sorry, I can't, but it
looks like this is already in good hands once an editor is assigned.

On Sat, 27 Jul 2019 at 02:21, Kyle Niemeyer <notifications@github.com>
wrote:

> Thanks @vsoch <https://github.com/vsoch> and @cole-brokamp
> <https://github.com/cole-brokamp>! We're still waiting to get an editor
> assigned, but I will definitely take you up on the offers to review.
> @cole-brokamp <https://github.com/cole-brokamp> you can certainly wait
> until you have completed your other review.
>
