### 2019-07-28 #1603 @whedon
[REVIEW]: containerit: Generating Dockerfiles for reproducible research with R
**Submitting author:** @nuest (<a href="http://orcid.org/0000-0002-0024-5046">Daniel Nüst</a>)
**Repository:** <a href="https://github.com/o2r-project/containerit/" target ="_blank">https://github.com/o2r-project/containerit/</a>
**Version:** 0.6.0
