### 2019-07-28 #1595 @danielskatz
👋 @terrytangyuan - you need to

otherwise, only the last assign is stored
