### 2019-07-28 #1595 @danielskatz
I opened https://github.com/openjournals/whedon-api/issues/63 for this - feel free to make more suggestions as to how this should better work
