### 2019-07-29 o2r-project/containerit#153 @vsoch
JoSS Review: Request for Proofreader
hey @nuest and @kyleniemeyer - I started reviewing your manuscript, but I keep hitting spelling mistakes, and weird sentences -  could you please find someone to proofread the draft that can help here? I'm happy to correct small issues, but I'm spending a huge amount of time with just grammar, and I'd like to ask that you find someone to read it over first, and I'll start my review after. Thank you!
