### 2019-07-29 #1603 @vsoch
## Review issues:
 - [ ] [Request for Proofreader](https://github.com/o2r-project/containerit/issues/153) the paper needs a native English speaker to read over first, there are many mistakes with grammar and sentence structure. I'll start my review after this. Thanks everyone!
