### 2019-07-29 o2r-project/containerit#153 @nuest
Hej @vsoch - I am not a native speaker, but have not had this feedback before. Thank you for pointing this out so directly, I'll organise a proofreading.
