### 2019-07-29 o2r-project/containerit#153 @vsoch
Thanks! Here is what I had edited so far (a few paragraphs in). You can likely ignore the changes of s -> z, but we should double check that JoSS doesn't favor one over the other.

# ContainerIt Review

## Manuscript

First I will review general comments for the manuscript

* I don't know if JoSS has a set language standard, but in case it's US, "standardised" should be "standardized"

> Running an analysis in a container increases trust in a workflow, as it can emulate an execution of the code independent from the author’s computer

I know what you are trying to say, but I don't think "trust" is the correct word. No container
can really be trusted, but rather, maybe "Running an analysis in a container increases reliability of a workflow?"

> as it can emulate an execution of the code independent from the author’s computer.

Emulate suggests that it's some sort of emulation. I think what you mean to say is something like "as it can execute code independently from the author's computer" or "as it can execute packaged code without relying on dependencies from the author's computer."

> However, especially domain scientists with limited programming experience struggle with the
complexity of capturing their computational environment in containers. containerit opens
up containerisation’s advantages to a much larger user base by assisting researchers in packag-
ing workflows based on R

I think you mean to have a comma after "environment in containers," and the sentence reads a little funny. I would also rephrase this to be something like:

> Since domain scientists with limited programming experience struggle with the
complexity of capturing their computational environment in containers, a tool like containerit
can help.

> Containerisation took off during the last years

This reads funny too. Maybe say "Containerization has recently taken off, or Containerization has taken off"

I don't think you need "during the last years," because you've referenced many recent
papers that give the reader a sense of the time period.

> The most widely used containerisation software is Docker
with the core building blocks image, which is (a) built from the instructions in a recipee called
Dockerfile, (b) executed as a container using a container runtime, and (c) moved between
systems as files (image tarballs) or based on an image registry (see Docker: Get Started).

This sentence also reads kind of funny, and note the typo that recipee only has one "e" at the end.

I think you want something like:

> The most widely used containerisation software is Docker, which builds container images from the instructions in a recipe called a Dockerfile, and assembles layers at runtime to create a container.

I don't think you need to add details about moving between systems and registries. If you want to keep that, the part of the sentence that needs to be reworked is

> The most widely used containerisation software is Docker with the core building blocks image

Recipe again

> The duality between recipee

recipe

> review and to some extend preservation

extent
