### 2019-07-29 #1603 @nuest
I'll organise a proofreading and will update both https://github.com/o2r-project/containerit/issues/153 and here.
