### 2019-07-29 #1595 @lheagy
Sorry for my late reply, I just returned from 4 days of backpacking and am getting caught up. If you are still in need of an editor here, I am happy to jump in and help, otherwise, @terrytangyuan it looks like you have things in motion!
