### 2019-07-29 o2r-project/containerit#153 @kyleniemeyer
Just jumping in here due to the JOSS connection—I think the specific suggestions @vsoch made look good and will improve readability. We do not really have a preference of s vs. z, other than consistency. (Though I am personally a stickler for Oxford commas and will fix those when they are missing...)

But, after reading through the paper (albeit fairly quickly), I didn't run into more than a few actual typos or errors—certainly not enough to impede my understanding. Most of the above suggestions are more about subtler word and phrasing choices that impede readability a bit. Good suggestions, yes, and I appreciate the time @vsoch put into providing them! But not really critical errors, in my opinion.

(Though I admit that I also generally struggle to get through something and focus on the content when I encounter issues like this myself 😄)

Also, as some friendly feedback, I think we should avoid any comments or questions about "native English" speakers/writers—if there are major issues with grammar, typos, or technical writing, then we can certainly point those out, but whether English is someone's first language does not have much to do with quality of technical writing, or ability to fix such.
