### 2019-07-29 2e658a0
Update paper.md
Adding commas after "e.g.", and ensured semicolon between grouped references
