### 2019-07-29 o2r-project/containerit#153 @kyleniemeyer
@nuest I made a few minor fixes (commas after "e.g.", ensuring semicolons between grouped references) here: https://github.com/o2r-project/containerit/pull/154
