### 2019-07-29 o2r-project/containerit#153 @vsoch
My apologies, in many journals I've submitted to / reviewed for there are explicit questions / boxes to check about if there are a large number of issues related to grammar. For example, the following sentences don't have correct structure:

 - The most widely used containerisation software is Docker with the core building blocks *image*,
 - Containerisation took off during **the last years**
 - However, **especially domain scientists** with limited programming experience struggle with the complexity of capturing their computational environment in containers. **containerit opens**

I wanted to suggest having the draft proofread before review. I'm okay with small grammar issues (commas and spelling mistakes) but as a reviewer my main focus should be the content and then the software.  I don't think it should be my responsibility to fix grammar in the draft, because that takes a really long time and it's a huge investment and extra burden. I started to do it, but stopped when I realized it's not what I should be reviewing. Again, I apologize if I offended someone with my request, I definitely didn't mean that.
