### 2019-07-29 o2r-project/containerit#153 @kyleniemeyer
Sorry, I think I tried to cram too many things into my message above and the end result was not clear.

I agree it isn't your (or any reviewer's) responsibility to edit or correct language in a submission. Suggesting the authors proofread / edit the paper to improve readability is absolutely appropriate for a reviewer, especially paired with a few examples as you did.

I appreciate the effort you did take in suggesting corrections, and I agree there are some additional places where things can be improved. Requesting another round of internal edits/proofreading is also perfectly fine, especially if you found the paper difficult to read. I personally didn't find the issues got in the way too much, but that was just my opinion—your opinion on this, especially as someone working closer to the application area, is very valuable!
