### 2019-07-29 o2r-project/containerit#153 @vsoch
Apologies again for not putting it very well - I edited the original comment so it's removed and we don't have to continually be offended by my lack of social skills :) I'm looking forward to reading the updated draft, let me know when to take a look!
