### 2019-08-08 52b624d
Update JOSS manuscript, closes #153
- feedback by @vsoch and @kyleniemeyer
- feedback by Celeste Brennecka, Science Writing Support Service (SWSS), University of Münster
