+++ paper.md
@@ -24,30 +24,33 @@ bibliography: paper.bib
 
 # Statement of Need
 
-[Linux containers](https://en.wikipedia.org/wiki/Operating-system-level_virtualization) have been demonstrated as a promising tool to increase transparency, portability, and reproducibility of research in several domains and use cases: data science [@boettiger_introduction_2015], software engineering research [@cito_using_2016], multi-step bioinformatics pipelines [@kim_bio-docklets_2017], standardised environments for exchangeable software [@belmann_bioboxes_2015], computational archeology [@marwick_computational_2017], packaging algorithms [@hosny_algorun_2016], or geographic object-based image analysis [@knoth_reproducibility_2017].
-Running an analysis in a container increases trust in a workflow, as it can emulate an execution of the code independent from the author's computer.
-However, especially domain scientists with limited programming experience struggle with the complexity of capturing their computational environment in containers.
-`containerit` opens up containerisation's advantages to a much larger user base by assisting researchers in packaging workflows based on R [@r_2018] in containers with user-friendly R commands.
+[Linux containers](https://en.wikipedia.org/wiki/Operating-system-level_virtualization) have become a promising tool to increase transparency, portability, and reproducibility of research in several domains and use cases: data science [@boettiger_introduction_2015], software engineering research [@cito_using_2016], multi-step bioinformatics pipelines [@kim_bio-docklets_2017], standardised environments for exchangeable software [@belmann_bioboxes_2015], computational archaeology [@marwick_computational_2017], packaging algorithms [@hosny_algorun_2016], or geographic object-based image analysis [@knoth_reproducibility_2017].
+Running an analysis in a container increases reliability of a workflow, as it can execute packaged code independently of the author's computer and its available configurations and dependencies.
+However, capturing a computational environment in containers can be complex, making container use difficult for domain scientists with limited programming experience.
+`containerit` opens up the advantages of containerisation to a much larger user base by assisting researchers in packaging workflows based on R [@r_2018] in containers with user-friendly R commands.
 
-Containerisation took off during the last years as a technology for packaging applications and their dependencies for fast, scalable, and secure [sandboxed](https://en.wikipedia.org/wiki/Sandbox_%28computer_security%29) deployments in cloud-based infrastructures [cf. @osnat_brief_2018].
-The most widely used containerisation software is [Docker](https://en.wikipedia.org/wiki/Docker_%28software%29) with the core building blocks _image_, which is (a) built from the instructions in a recipee called `Dockerfile`, (b) executed as a _container_ using a _container runtime_, and (c) moved between systems as files (image tarballs) or based on an _image registry_ (see [Docker: Get Started](https://docs.docker.com/get-started/)).
+Recently containerisation took off as a technology for packaging applications and their dependencies for fast, scalable, and secure [sandboxed](https://en.wikipedia.org/wiki/Sandbox_%28computer_security%29) deployments in cloud-based infrastructures [cf. @osnat_brief_2018].
+The most widely used containerisation software is [Docker](https://en.wikipedia.org/wiki/Docker_%28software%29) with the following core building blocks (cf. [Docker: Get Started](https://docs.docker.com/get-started/)):
+The _image_ is built from the instructions in a recipe called `Dockerfile`.
+The image is executed as a _container_ using a _container runtime_.
+An image can be moved between systems as a file (image tarball) or based on an _image registry_.
 A `Dockerfile` may use the image created by another `Dockerfile` as the starting point, a so-called _base image_.
-While manual alteration of containers is possible, the common practice is to conduct all configuration with scripts and instructions originating in the `Dockerfile`.
[…]
