### 2019-08-08 o2r-project/containerit#153 @nuest
Thank both of you for the helpful comments. Based on your feedback and feedback provided by a native speaker from my university's writing support service (a stupid oversight on my side not to engage with them earlier) I have committed a revised manuscript.

Re. BE/AE: I use British English throughout the document and hope nothing has slipped by so it is consistent.

Here are the changes: https://github.com/o2r-project/containerit/commit/52b624dad4d880967c061ebe257322b292806ffd
