+++ README.Rmd
@@ -17,6 +17,7 @@ knitr::opts_chunk$set(
 The packages's website at [https://o2r.info/containerit/](https://o2r.info/containerit/).
 A good summary of what the package does can be found in this [o2r project blog post](http://o2r.info/2017/05/30/containerit-package/).
 
+[![status](http://joss.theoj.org/papers/6f3b6a7cca0fa133966f13f260f0360d/status.svg)](http://joss.theoj.org/papers/6f3b6a7cca0fa133966f13f260f0360d)
 [![Project Status: WIP - Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](http://www.repostatus.org/badges/latest/wip.svg)](http://www.repostatus.org/#wip)
 [![Travis CI build status](https://travis-ci.org/o2r-project/containerit.svg?branch=master)](https://travis-ci.org/o2r-project/containerit)
 [![Appveyor build status](https://ci.appveyor.com/api/projects/status/2242hcwagoafxaxq?svg=true)](https://ci.appveyor.com/project/nuest/containerit-rrvpq)
