### 2019-08-08 o2r-project/containerit#155 @vsoch
JoSS Review 1: Manuscript
## Manuscript
This is for https://github.com/openjournals/joss-reviews/issues/1603, and details a review of the manuscript.

Overall, huge improvement from #153 ! The introduction opens with a short background of containers, challenges, and why a tool like containerit would be needed. This is spot on to what I'd want to see - I'm one paragraph in and I have a clear understanding of the issue at hand, and how the software described in the paper can help.

The following points are tiny (and some just commentary that you might find interesting).

> An important advantage of containers over virtual machines is that their duality between
recipee

"Recipe" should have just one e - "recipe"

> Even if an image cannot be executed or a Dockerfile can no longer be built, the instructions in the Dockerfile are human-readable, and files in the image can be extracted to recreate an environment that closely resembles the original.

I really like this point - I've never considered it like this! It's typical to think that if a container
doesn't build, you're done, but you're totally right - having a recipe means you can try to restore
functionality by upgrading versions, trying different bases, etc.

> To enable and leverage the stakeholders’ infrastructure, container creation must become easier and more widespread.

Side note (for your own interest) if you are interested in long term survival of container layers (or single binary), tar isn't the beastie that you seek (there is a good post here https://twitter.com/lordcyphar/status/1087138958767804418 but the author's site seems down at the moment).  A single, read only binary like SIF (less suseptible to bit rot) is a better unit for long term survival of a container -> https://archive.sylabs.io/2018/03/sif-containing-your-containers/. The good news is that any Docker container can be converted - it just comes down to dumping layers into a root to build a SIF binary.

But if you do decide to try converting any of the rock-r images to Singularity, take absolute precaution because the init scripts in the Docker container can [hugely mess up](https://github.com/sylabs/singularity/issues/3777) your machine! What you can do instead is remove the complicated start up process, and just run the software / services that you need [here is an example with rstudio](https://github.com/singularityhub/singularity-compose-examples/blob/master/rstudio-simple/rstudio/Singularity). Notice the startscript (akin to the entrypoint) is just starting the server :) I'm getting distracted - back to review!

Holy cow! You referenced the Experiment Factory! Totally awesome! :D To be fair, it's an entirely different use case - it isn't to reproduce an analysis, but generates an entire web application to serve experiments. You are correct that to build the container off the bat, one extra file is added (the startscript).

> bespoke

Just a nitpick - you use this three times in the paper, and it's noticeable. What about choosing a different word for one or two so it doesn't stand out?
