### 2019-08-08 o2r-project/containerit#156 @vsoch
JoSS Review 2: Repository and Install
### Repository

I can confirm that the source code is available at https://github.com/o2r-project/containerit/, and the VERSION is correct. The authorship is appropriate, and an OSI approved LICENSE is included.

#### Contributing

The author should consider adding a CONTRIBUTING.md, which can be in the root of the respository or a .github folder. The [Contributor's Covenant](https://www.contributor-covenant.org/) is popular, but others are available (you can do a search online for templates, etc.) It's good practice to have sections in the README that describe broadly how to contribute (and link to that file) along with how to ask for support (questions, issues, etc.)

#### Installation

The steps are very nicely presented in the README! I'm not a super active R user, so I don't have the remotes::install_github command (how to load it from a library) in my head. Could this be provided?

```R
# install.packages("remotes")
remotes::install_github("o2r-project/containerit")
```

Just a note that I hit this error - my version of R is old enough so it would be expected to use bioconductor and not BiocManager:

```R
Error: Failed to install 'containerit' from GitHub:
  (converted from warning) package ‘BiocManager’ is not available (for R version 3.4.4)
```

I decided to instead install in a Docker image:

```bash
docker run --rm -p 8787:8787 -e PASSWORD=meatballs rocker/rstudio
```

That seems to start to work, but fails with:

```bash
trying URL 'https://mran.microsoft.com/snapshot/2019-08-06/src/contrib/digest_0.6.20.tar.gz'
Error in download.file(url, destfile, method, mode = "wb", ...) :
  (converted from warning) URL 'https://mran.microsoft.com/snapshot/2019-08-06/src/contrib/digest_0.6.20.tar.gz': status was 'Couldn't connect to server'
Error: Failed to install 'containerit' from GitHub:
  (converted from warning) download of package ‘digest’ failed
```

Can you advise? If I've hit two different errors trying to install, others might too.
