### 2019-08-08 o2r-project/containerit#156 @vsoch
okay this seems to work!
```R
install.packages('digest', repos='http://cran.us.r-project.org')
```
Let me see if I can create a Dockerfile that could be provided to the user to bypass these issues. Even if the user winds up eventually installing on their host, we want to make their first test / introduction to the software as painless as possible.
