### 2019-08-08 o2r-project/containerit#156 @vsoch
okay! So this is a bit of a catch-22 - I was able to create a Dockerfile that would allow me to install containerit (sorry it took me so long, literally was try and fail all day!)

```dockerfile
FROM rocker/rstudio

# docker build -t containerit .
# docker run --rm -p 8787:8787 -e PASSWORD=meatballs containerit
# open to localhost:8787, username rstudio and password set above

RUN apt-get update && apt-get install -y zlib1g-dev libxml2-dev && \
    Rscript -e "install.packages('digest', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('debugme', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('BiocManager', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('remotes')" && \
    Rscript -e "install.packages('formatR', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('stringr', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('purrr', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('R6', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('devtools', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('plogr', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('utf8', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('semver', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('automagic', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('shiny', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('rjson', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('futile.logger', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('miniUI', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('shinyFiles', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('stevedore', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('versions', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('usethis', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('git2r', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('roxygen2', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('lambda.r', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('futile.options', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('htmltools', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('httpuv', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('xtable', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('sourcetools', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('later', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('promises', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('xml2', repos='http://cran.us.r-project.org')" && \
    Rscript -e "install.packages('devtools')" && \
    Rscript -e "remotes::install_github('o2r-project/containerit')"
```
And then I was able to start the container (see the header) and generate a recipe, as per the instructions in the readme (well done here!) And the generated recipe is the following:

```dockerfile
FROM rocker/r-ver:3.6.1
LABEL maintainer="rstudio"
RUN export DEBIAN_FRONTEND=noninteractive; apt-get -y update \
  && apt-get install -y libcurl4-openssl-dev \
	libssl-dev \
	make
RUN ["install2.r", "assertthat", "backports", "crayon", "curl", "desc", "digest", "formatR", "fs", "futile.logger", "futile.options", "htmltools", "httpuv", "jsonlite", "lambda.r", "later", "magrittr", "mime", "miniUI", "pillar", "pkgconfig", "promises", "R6", "Rcpp", "rlang", "rprojroot", "rstudioapi", "semver", "shiny", "shinyFiles", "stevedore", "stringi", "stringr", "tibble", "versions", "xtable"]
WORKDIR /payload/
CMD ["R"]
```
I was really happy because the above is much cleaner / nicer looking than my Dockerfile, and also very funny because I used containerit to fix a dependency issue with containerit! What do you think about adding this Dockerfile to the repository? I'll open a PR for it soon.
