### 2019-08-08 70947b0
adding basic Dockerfile for rstudio
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
