### 2019-08-09 ccf524a
adding missing dependencies
Signed-off-by: Vanessa Sochat <vsochat@stanford.edu>
