### 2019-08-09 o2r-project/containerit#156 @vsoch
See https://github.com/o2r-project/containerit/pull/157 - the recipe is working and provides an easy way to quickly get started! I'd recommend an automated build from the repository so the user doesn't have to wait for it to build (it takes a bit of time) but instead can just pull from Docker Hub.
