### 2019-08-14 #1603 @terrytangyuan
👋 @vsoch @cole-brokamp Looks like the paper has been revised. Would you be able to start the review soon?
