### 2019-08-14 #1603 @vsoch
And actually I forgot to link this one – sorry about that! https://github.com/o2r-project/containerit/issues/156
