### 2019-08-14 o2r-project/containerit#156 @cole-brokamp
>
> Just a note that I hit this error - my version of R is old enough so it would be expected to use bioconductor and not BiocManager:
>
> ```r
> Error: Failed to install 'containerit' from GitHub:
>   (converted from warning) package ‘BiocManager’ is not available (for R version 3.4.4)
> ```

This package requires R 3.5.0 or greater.  This is probably why you're seeing the install error for its dependency BiocManager not being available?

https://github.com/o2r-project/containerit/blob/de6749f943a0a20bffadb285470afb205759202c/DESCRIPTION#L47-L48
