### 2019-08-14 o2r-project/containerit#156 @vsoch
I hit that exact error @cole-brokamp ! That's why I created the Docker container that would give you a quick start - it's in the pull request here --> https://github.com/o2r-project/containerit/pull/157

It's important that a user (like yourself) is able to try it out quickly - if there is high barrier to that, it can hugely hurt a piece of software from being able to take off.

I'm not sure about the status of the tests, because I didn't change anything but (for the other PR too) they seem to be failing. @nuest must be busy because he hasn't had a chance to comment on either, so I paused my review until he's back.
