### 2019-08-14 o2r-project/containerit#159 @cole-brokamp
JoSS review (reviewer #2)
Hello, opening up an issue separately from @vsoch's here for some of my more R specific questions/comments.  I'll try not to duplicate anything that she has already brought up.

#### Installation

I also ran into issue installing the package. `remotes::install_github()` returned an error:

```r
Error: HTTP error 422.
  No commit found for SHA: gaborcsardi/gh
```

but using `pak::pkg_install()` worked just fine.  We need to either determine the root cause of this install failure or provide user with alternative install instructions.  I wonder if its because pak doesn't install suggested dependencies unless explicitly directed to?

#### Statement of Need

The first sentence of the readme does state what the software does, but it might be helpful to also identify targeted users... maybe R users who are unfamiliar with creating Dockerfiles and containers?

#### Functionality Documentation

Trying to run the example in the README seems to work for generating the dockerfile object, but printing it to the console does not result in what is in the README.  Instead, I get what appears to be a standard printing of the S6 object's slots:

```r
An object of class "Dockerfile"
Slot "image":
An object of class "From"
Slot "image":
[1] "rocker/r-ver"

Slot "postfix":
An object of class "Tag"
[1] "3.5.3"

Slot "maintainer":
An object of class "Label"
Slot "data":
$maintainer
[1] "cole"

Slot "multi_line":
[1] FALSE

Slot "instructions":
[[1]]
An object of class "Run_shell"
Slot "commands":
[1] "export DEBIAN_FRONTEND=noninteractive; apt-get -y update"
[2] "apt-get install -y git-core \\\n\tlibcurl4-openssl-dev \\\n\tlibssl-dev \\\n\tlibxml2-dev \\\n\tmake"

[[2]]
An object of class "Run"
Slot "exec":
[1] "install2.r"

Slot "params":
 [1] "assertthat"     "backports"      "broom"          "callr"          "cellranger"     "cli"            "codetools"      "colorspace"     "crayon"
[10] "curl"           "desc"           "digest"         "dplyr"          "forcats"        "formatR"        "fs"             "futile.logger"  "futile.options"
[19] "generics"       "ggplot2"        "glue"           "gtable"         "haven"          "hms"            "htmltools"      "httpuv"         "httr"
[28] "jsonlite"       "lambda.r"       "later"          "lattice"        "lazyeval"       "lintr"          "lubridate"      "magrittr"       "mime"
[37] "miniUI"         "modelr"         "munsell"        "nlme"           "pillar"         "pkgbuild"       "pkgconfig"      "plyr"           "prettyunits"
[46] "processx"       "promises"       "ps"             "purrr"          "R6"             "Rcpp"           "readr"          "readxl"         "remotes"
[55] "rex"            "rlang"          "rprojroot"      "rstudioapi"     "rvest"          "scales"         "semver"         "shiny"          "shinyFiles"
[64] "stevedore"      "stringi"        "stringr"        "tibble"         "tidyr"          "tidyselect"     "tidyverse"      "vctrs"          "versions"
[73] "withr"          "xml2"           "xtable"         "zeallot"

[[3]]
An object of class "Run"
Slot "exec":
[1] "installGithub.r"

Slot "params":
[1] "cole-brokamp/CB@7ea91855c70c9586ee9546ecbdb10b044f3ef100"    "hadley/memoise@1650ad7f9c27d2e0a932e50b16eedb574e8050df"
[3] "jalvesaq/colorout@cc5fbfa1f165bd4762588b463157a2688b116c5f"  "jimhester/highlite@767b122ef47a60a01e1707e4093cf3635a99c86b"
[5] "jimhester/lookup@e1e94d334e55660709d995f651bc746f7b7be4f1"   "r-lib/pak@7a69eaa18b7a250d044e9c6e948ff8930823a314"

[[4]]
An object of class "Workdir"
Slot "path":
[1] "/payload/"

Slot "entrypoint":
NULL

Slot "cmd":
An object of class "Cmd"
Slot "exec":
[1] "R"

Slot "params":
[1] NA

Slot "form":
[1] "exec"
```

Hopefully this isn't something wrong with my installation, but no matter the cause, the core functionality doesn't seem to proceed as advertised in this case.

#### Manuscript

- you may want to consider listing ![renv](https://github.com/rstudio/renv) as a another related work
- in Related Work section: "namerly" -> "namely"
