### 2019-08-14 #1603 @cole-brokamp
Also just added some of my own issues and comments as a part of my review over here: https://github.com/o2r-project/containerit/issues/159
