### 2019-08-14 o2r-project/containerit#159 @cole-brokamp
As suggested by @vsoch here: https://github.com/o2r-project/containerit/issues/156, it would be great to have instructions for using a docker container to try out the package without having to install R and the long list of dependencies.  Specifically, adding a part of your vignette (https://o2r.info/containerit/articles/container.html) to the README would help new users try it out easier and quicker!
