### 2019-08-15 o2r-project/containerit#155 @nuest
_Thank you for the feedback!_

And especially thank you for the pointers about problems with tarballs - some great food for thought there! I've not come across this in the handful of papers I've found working on preservation of containers, so I've taken a few notes but will surely revisit those posts a few times.

_Re. R Singularity images_: I've made some experiments a while back, but have not checked for a long time, and it seems there are a bunch of images doing something with RStudio (via https://singularity-hub.org/search). I've taken a note for a paper idea here: https://github.com/nuest/rockerverse-paper/issues/6

At a first glance, the existing images mostly bootstrap Rocker images or just install R. I'm getting distracted, too :-) _Are your aware of any efforts to consolidate the different R images?_ Not to piggyback this issue_ your feedback at https://github.com/nuest/rockerverse-paper/issues/6 (or [contribution](https://github.com/nuest/rockerverse-paper/issues/3)?) would be very welcome.
