### 2019-08-15 o2r-project/containerit#155 @vsoch
For Singularity, (and I run Singularity Hub, so I can answer this with certainty) the images there are primarily done by individuals that want to run a container-based analysis on HPC, so that's why you don't see any sort of library of R base images (it's not the direct goal of Singularity Hub and no user has the incentive to provide it). If you were wanting an equivalent base for R images, you could try engaging with the Sylabs folks, their [cloud](https://cloud.sylabs.io/library) does provide several base images.

Ah! Just looking at the link - I'll add comments there!
