### 2019-08-15 4c32a7d
minor wording improvements, fix typo - closes #155
getting rid of one of my favourite words with more suitable ones!
