### 2019-08-15 1b73a6c
mention images in README
as suggested in #156
