### 2019-08-15 59fb245
more explicitly describe target users and mention renv (see #159)
