### 2019-08-15 o2r-project/containerit#159 @nuest
I've started a PR to track the changes after the first round of review: https://github.com/o2r-project/containerit/pull/160

Re. the printing of the object rather than the function: _How did you install the package?_ [There is a function](https://github.com/o2r-project/containerit/blob/master/R/Class-Dockerfile.R#L54) that should print the Dockerfile and not the object, but maybe it was not properly registered in your installation with `pak`? I will investigate.

Re. users in the statement of need: extended the last sentence of the first paragraph to make that a bit more explicit. _Better?_
