### 2019-08-15 o2r-project/containerit#156 @nuest
@vsoch Thanks for the first round of feedback. I'll get to your suggestion of a user Dockerfile in the PR.

I've started a PR to track the changes of the first round of reviews: https://github.com/o2r-project/containerit/pull/160

- added a "Prerequisites" section to README
- added comment with installation command for `remotes`
- a code of conduct is in `CONDUCT.md`, also mentioned in the readme
- clarified in the `README` how the existing Docker images can be used to run `RStudio`: https://github.com/nuest/containerit/commit/1b73a6cf939e5ebfb2bc2d74198288d4108c0094

------

- will add a `CONTRIBUTING.md` about the contribution process (probably a few words about fork and pull, tag some issues as good "first issues")
