### 2019-08-15 o2r-project/containerit#156 @vsoch
Awesome!  Your changes address all of my concerns, minus the last one with the CONTRIBUTING.md. Ping me here when you've added and we can close the issue.
