### 2019-08-16 e16adaf
add CONTRIBUTING.md
partial of #156
