+++ README.Rmd
@@ -22,6 +22,7 @@ A good summary of what the package does can be found in this [o2r project blog p
 [![Travis CI build status](https://travis-ci.org/o2r-project/containerit.svg?branch=master)](https://travis-ci.org/o2r-project/containerit)
 [![Appveyor build status](https://ci.appveyor.com/api/projects/status/2242hcwagoafxaxq?svg=true)](https://ci.appveyor.com/project/nuest/containerit-rrvpq)
 [![](https://www.r-pkg.org/badges/version/containerit)](#)
+[![Join the chat at https://gitter.im/o2r-project/containerit](https://badges.gitter.im/o2r-project/containerit.svg)](https://gitter.im/o2r-project/containerit)
 <!-- https://www.r-pkg.org/pkg/containerit -->
 
 ![containerit logo](inst/logo.png)
@@ -131,10 +132,12 @@ Courtesy of a great contribution by a [group of enthusiasts](https://github.com/
 
 ## Contribute
 
-All help is welcome: asking questions, providing documentation, testing, or even programming.
+All help is welcome: asking questions, providing documentation, testing, or even development.
 
 Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md).
 By participating in this project you agree to abide by its terms.
+
+See [CONTRIBUTING.md](CONTRIBUTING.md) for details.
 
 ## Development
 
