### 2019-08-16 o2r-project/containerit#156 @nuest
Contribution documentation added @vsoch : https://github.com/nuest/containerit/commit/e16adaf037a5d95e2dd2a8dcde59173fd75b193c
