### 2019-08-16 08e1504
try to clarify target user, partial of #159
