### 2019-08-16 o2r-project/containerit#159 @nuest
I've tried installation with `pak` in a `rocker/geospatial` container, and had the same result when I did not load the package via `library(..)` but only used it via `containerit::..`:

```
daniel@nuest:~$ docker run --rm -it rocker/geospatial R
> install.packages("pak")
> pak::pkg_install(pkg = "o2r-project/containerit")
ℹ Checking for package metadata updates
✔ Downloaded metadata files, 1.6 MB in 6 files.
✔ Updating metadata database
✔ Using cached package metadata

→ Will install 12 packages:
  o2r-project/containerit, automagic, futile.logger, futile.options, lambda.r,
  rjson, semver, shinyFiles, stevedore, versions, github::r-hub/sysreqs,
  debugme

→ Will not update 69 packages.

→ Will download 10 CRAN packages (2.96 MB).
→ Will download 2 packages with unknown size.

? Do you want to continue (Y/n) y
ℹ Building automagic 0.5.1
ℹ Building futile.options 1.0.1
ℹ Building lambda.r 1.2.3
ℹ Building rjson 0.2.20
ℹ Building semver 0.2.0
ℹ Building shinyFiles 0.7.3
ℹ Building stevedore 0.9.1
ℹ Building versions 0.3
✔ Built automagic 0.5.1 [3.1s]                                                  ..
ℹ Building debugme 1.1.0                                                        ..
✔ Built futile.options 1.0.1 [3.4s]                                             ..
✔ Installed automagic 0.5.1  [88ms]                                             ..
✔ Installed futile.options 1.0.1  [112ms]                                       ..
✔ Built versions 0.3 [3.6s]                                                     ..
✔ Installed versions 0.3  [64ms]                                                ..
✔ Built lambda.r 1.2.3 [5.6s]                                                   ..
✔ Installed lambda.r 1.2.3  [65ms]                                              ..
ℹ Building futile.logger 1.4.3                                                  ..
✔ Built rjson 0.2.20 [5.9s]                                                     ..
✔ Installed rjson 0.2.20  [86ms]                                                ..
✔ Built debugme 1.1.0 [3.5s]                                                    ..
✔ Installed debugme 1.1.0  [130ms]                                              ..
ℹ Building sysreqs 1.0.0.9000                                                   ..
✔ Built shinyFiles 0.7.3 [7.4s]                                                 ..
✔ Installed shinyFiles 0.7.3  [92ms]                                            ..
✔ Built futile.logger 1.4.3 [3.5s]                                              ..
✔ Built sysreqs 1.0.0.9000 [2.6s]                                               ..
✔ Installed futile.logger 1.4.3  [101ms]                                        ..
✔ Installed sysreqs 1.0.0.9000 (github::r-hub/sysreqs@3860f2b) [37ms]           ..
✔ Built stevedore 0.9.1 [10.8s]                                                 ..
✔ Installed stevedore 0.9.1  [47ms]                                             ..
✔ Built semver 0.2.0 [16.5s]                                                    ..
✔ Installed semver 0.2.0  [70ms]
ℹ Building containerit 0.5.0
✔ Built containerit 0.5.0 [5.5s]
✔ Installed containerit 0.5.0 (github::o2r-project/containerit@de6749f) [48ms]
✔ 1 + 80 pkgs | kept 69, updated 0, new 12 | downloaded 12 (3.18 MB) [42s]
> df <- containerit::dockerfile()
INFO [2019-08-16 07:57:41] Going online? TRUE  ... to retrieve system dependencies (sysreq-api)
INFO [2019-08-16 07:57:41] Trying to determine system requirements for the package(s) 'assertthat,backports,callr,crayon,curl,desc,digest,filelock,formatR,fs,futile.logger,futile.options,htmltools,httpuv,jsonlite,lambda.r,later,magrittr,mime,miniUI,pak,pillar,pkgconfig,processx,promises,ps,R6,Rcpp,rlang,rprojroot,semver,shiny,shinyFiles,stevedore,stringi,stringr,tibble,versions,xtable' from sysreqs online DB
INFO [2019-08-16 07:57:59] Adding CRAN packages: assertthat, backports, callr, crayon, curl, desc, digest, filelock, formatR, fs, futile.logger, futile.options, htmltools, httpuv, jsonlite, lambda.r, later, magrittr, mime, miniUI, pak, pillar, pkgconfig, processx, promises, ps, R6, Rcpp, rlang, rprojroot, semver, shiny, shinyFiles, stevedore, stringi, stringr, tibble, versions, xtable
INFO [2019-08-16 07:57:59] Created Dockerfile-Object based on sessionInfo
> print(df)
An object of class "Dockerfile"
Slot "image":
An object of class "From"
Slot "image":
[1] "rocker/r-ver"

Slot "postfix":
An object of class "Tag"
[1] "3.6.0"

Slot "maintainer":
An object of class "Label"
Slot "data":
$maintainer
[1] "root"

Slot "multi_line":
[1] FALSE

Slot "instructions":
[[1]]
An object of class "Run_shell"
Slot "commands":
[1] "export DEBIAN_FRONTEND=noninteractive; apt-get -y update"
[2] "apt-get install -y libcurl4-openssl-dev \\\n\tlibssl-dev \\\n\tmake"

[[2]]
An object of class "Run"
Slot "exec":
[1] "install2.r"

Slot "params":
 [1] "assertthat"     "backports"      "callr"          "crayon"
 [5] "curl"           "desc"           "digest"         "filelock"
 [9] "formatR"        "fs"             "futile.logger"  "futile.options"
[13] "htmltools"      "httpuv"         "jsonlite"       "lambda.r"
[17] "later"          "magrittr"       "mime"           "miniUI"
[21] "pak"            "pillar"         "pkgconfig"      "processx"
[25] "promises"       "ps"             "R6"             "Rcpp"
[29] "rlang"          "rprojroot"      "semver"         "shiny"
[33] "shinyFiles"     "stevedore"      "stringi"        "stringr"
[37] "tibble"         "versions"       "xtable"

[[3]]
An object of class "Workdir"
Slot "path":
[1] "/payload/"

Slot "entrypoint":
NULL

Slot "cmd":
An object of class "Cmd"
Slot "exec":
[1] "R"

Slot "params":
[1] NA

Slot "form":
[1] "exec"

> library("containerit")
> print(df)
FROM rocker/r-ver:3.6.0
LABEL maintainer="root"
RUN export DEBIAN_FRONTEND=noninteractive; apt-get -y update \
  && apt-get install -y libcurl4-openssl-dev \
	libssl-dev \
	make
RUN ["install2.r", "assertthat", "backports", "callr", "crayon", "curl", "desc", "digest", "filelock", "formatR", "fs", "futile.logger", "futile.options", "htmltools", "httpuv", "jsonlite", "lambda.r", "later", "magrittr", "mime", "miniUI", "pak", "pillar", "pkgconfig", "processx", "promises", "ps", "R6", "Rcpp", "rlang", "rprojroot", "semver", "shiny", "shinyFiles", "stevedore", "stringi", "stringr", "tibble", "versions", "xtable"]
WORKDIR /payload/
CMD ["R"]

```

Can you confirm that behaviour?
