### 2019-08-16 775cae5
Add minimal Dockerfile and a test for it, see also #157
