### 2019-08-16 1cb39e6
fix some tests
