### 2019-08-16 fa76912
warnings are not errors on appveyor
