+++ appveyor.yml
@@ -46,7 +46,7 @@ environment:
 
   matrix:
     - R_VERSION: release
-      WARNINGS_ARE_ERRORS: 1
+      WARNINGS_ARE_ERRORS: 0
     #- R_VERSION: devel
     #  WARNINGS_ARE_ERRORS: ""
 
