### 2019-08-16 #1603 @nuest
I've collected my updates based on reviewer feedback in this PR: https://github.com/o2r-project/containerit/pull/160 (Tests also work again).

I suggest to merge that PR and update the tag as soon as the reviewers confirm everything mention so far is addressed.

@cole-brokamp The only open thread AFAICS is the problem that the printing did not work properly during your tests: https://github.com/o2r-project/containerit/issues/159#issuecomment-521924165 It would be great if you could try again.
