### 2019-08-16 o2r-project/containerit#156 @vsoch
Great work! My questions and issues have been addressed here, closing issue.
