### 2019-08-16 #1603 @vsoch
@nuest now that testing is fixed and you've made good changes to the docs and paper, when the pull request is merged (and I've checked the version) I can finish up the review. Thanks for your hard work on this!
