### 2019-08-16 o2r-project/containerit#159 @cole-brokamp
I think this is a moot point because my problems with `remotes::install_github('o2r-project/containerit')` disappeared when using a fresh R version in docker (with either MRAN or CRAN used to support the install).  So I'm going to chalk this up to a problem with one of the dependencies in my local library.  Things proceed as expected now.
