### 2019-08-16 o2r-project/containerit#159 @cole-brokamp
The RStudio addins are so very cool!!

You can take or leave my suggestions below about the manuscript, but everything for the review portion looks complete.  Thanks!

> #### Manuscript
>
>     * you may want to consider listing ![renv](https://github.com/rstudio/renv) as a another related work
>
>     * in Related Work section: "namerly" -> "namely"
