### 2019-08-16 #1603 @cole-brokamp
Everything looking good to me.  I'm all checked off.  Thanks!
