### 2019-08-20 #1603 @nuest
:tada: Thank you @vsoch and @cole-brokamp for the great review experience and the useful feedback, especially on improving the manuscript and increasing accessibility for newcomers and potential contributors.

https://github.com/o2r-project/containerit/pull/160 is merged.
