### 2019-08-20 #1603 @nuest
I also updated the tag: https://github.com/o2r-project/containerit/releases/tag/v0.5.0

_Is it time now to create the release on Zenodo?_
