### 2019-08-20 21db07f
fix pkgdown landing page (with logo) and use https wherever possible
