+++ vignettes/container.Rmd
@@ -69,7 +69,7 @@ rm -r /tmp/containerit
 
 ```{bash script, eval=FALSE}
 mkdir /tmp/containerit
-echo "install.packages('fortunes', repos = 'http://cran.r-project.org'); fortunes::fortune();" >> /tmp/containerit/script.R
+echo "install.packages('fortunes', repos = 'https://cran.r-project.org'); fortunes::fortune();" >> /tmp/containerit/script.R
 
 docker run --rm -v /tmp/containerit:/data o2rproject/containerit R -q -e "setwd('/data'); scriptFile <- 'script.R'; write(dockerfile(from = scriptFile, cmd = CMD_Rscript(basename(scriptFile))));"
 
