+++ vignettes/basic.Rmd
@@ -114,7 +114,7 @@ clear_all()
 ### Create Dockerfile from script
 
 This example uses the [rgdal](https://CRAN.R-project.org/package=rgdal) package because it has system library dependencies, namely GDAL and PROJ.
-Code snippets are taken from the [sp gallery](http://edzer.github.io/sp/).
+Code snippets are taken from the [sp gallery](https://edzer.github.io/sp/).
 Here is some regular R code loading a file and plotting it.
 
 ```{r example_geo, eval=FALSE}
