### 2019-08-20 #1603 @nuest
FYI: I'm currently fixing some small glitches since an online resource used in the tests is not available right now. Hardening the tests should do it.
