### 2019-08-20 f75510f
skip tests depending on crandeps.r-pkg.org if not available, closes #163
