+++ tests/testthat/test_find_systemrequirements.R
@@ -2,7 +2,15 @@
 
 context("Find system requrirements")
 
+skip_if_crandeps_offline <- function() {
+  skip_if_not_installed("httr")
+  if (httr::status_code(httr::HEAD(url = "crandeps.r-pkg.org")) != 200)
+    skip("crandeps not available")
+}
+
 test_that("System requirements for sp can be detected OFFLINE and soft", {
+  skip_if_crandeps_offline()
+
   output <- capture_output(
     deps <- containerit:::.find_system_dependencies("sp",
                                       platform = containerit:::.debian_platform,
@@ -15,6 +23,8 @@ test_that("System requirements for sp can be detected OFFLINE and soft", {
 })
 
 test_that("System requirements for sp can be detected OFFLINE and unsoft", {
+  skip_if_crandeps_offline()
+
   output <- capture_output(
     deps <- containerit:::.find_system_dependencies("sp",
                                                     platform = containerit:::.debian_platform,
@@ -25,6 +35,8 @@ test_that("System requirements for sp can be detected OFFLINE and unsoft", {
 })
 
 test_that("System requirements for rgdal can be detected OFFLINE", {
+  skip_if_crandeps_offline()
+
   output <- capture_output(
     deps <- containerit:::.find_system_dependencies("rgdal",
                                                     platform = containerit:::.debian_platform,
