### 2019-08-20 f980627
fix some tests around GitHub install and packaging DESCRIPTION files
installGithub.r needs {devtools} before R 3.3.0, so let's only test R >= 3.4 and make double sure {remotes} is available and the image can be built and run
