### 2019-08-20 #1603 @nuest
Everything's in order now (https://github.com/o2r-project/containerit/releases/tag/v0.5.0). Also fixed a DOI URL, so...
