### 2019-08-20 #1603 @vsoch
@nuest has addressed all of my issues! Here are the first two that I forgot to link:

 - https://github.com/o2r-project/containerit/issues/153
 - https://github.com/o2r-project/containerit/issues/156

The last little bit would be to bump the version, in the case that the software was tweaked for review? That's my last bullet, so when we've verified the version, and released on Zenodo, I recommend this software for publication in JoSS. Great job addressing all of the concerns @nuest!
