### 2019-08-20 #1603 @terrytangyuan
Thanks everyone! @nuest does v0.5.0 contain all the fixes you want or do you still need to release another one?
