+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: containerit
 Title: Package R in Docker Containers
-Version: 0.5.0
+Version: 0.6.0
 Authors@R: 
     c(person(given = "Daniel",
              family = "Nuest",
