### 2019-08-20 #1603 @nuest
Version `0.6.0` is released on GitHub: https://github.com/o2r-project/containerit/releases/tag/v0.6.0

May I release to Zenodo?
