### 2019-08-20 #1603 @terrytangyuan
Yes @nuest. Please make sure the authors field and version in Zenondo match with the paper.
