### 2019-08-21 93005e0
Update paper.md
Minor copy editing
