+++ paper.bib
@@ -1,5 +1,5 @@
 @Manual{allaire_rmarkdown_2018,
-  title = {rmarkdown: Dynamic Documents for R},
+  title = {{rmarkdown}: Dynamic Documents for {R}},
   author = {JJ Allaire and Yihui Xie and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang and Richard Iannone},
   year = {2018},
   note = {R package version 1.11},
@@ -56,7 +56,7 @@
 }
 
 @Manual{brokamp_automagic_2017,
-  title = {automagic: Automagically Document and Install Packages Necessary to Run R
+  title = {automagic: Automagically Document and Install Packages Necessary to Run {R}
 Code},
   author = {Cole Brokamp},
   year = {2017},
@@ -65,7 +65,7 @@ Code},
 }
 
 @inproceedings{ChirigatiRSF16,
-  title = {ReproZip: Computational Reproducibility With Ease},
+  title = {{ReproZip}: Computational Reproducibility With Ease},
   author = {Chirigati, Fernando and Rampin, R{\'e}mi and Shasha, Dennis and Freire, Juliana},
   year = {2016},
   isbn = {978-1-4503-3531-7},
@@ -100,7 +100,7 @@ Code},
 }
 
 @Manual{csardi_callr_2018,
-  title = {callr: Call R from R},
+  title = {callr: Call {R} from {R}},
   author = {Gábor Csárdi and Winston Chang},
   year = {2018},
   note = {R package version 3.1.1},
@@ -119,7 +119,7 @@ Code},
 }
 
 @Manual{fay_dockerfiler_2018,
-  title = {dockerfiler: Easy Dockerfile Creation from R},
+  title = {dockerfiler: Easy {Dockerfile} Creation from {R}},
   author = {Colin Fay},
   year = {2018},
   note = {R package version 0.1.1},
@@ -127,7 +127,7 @@ Code},
 }
 
 @Manual{fitzjohn_stevedore_2019,
-  title = {stevedore: Docker Client},
+  title = {stevedore: {Docker} Client},
   author = {Rich FitzJohn},
   year = {2019},
   note = {R package version 0.9.1},
@@ -166,7 +166,7 @@ Code},
   doi = {10.5063/schema/codemeta-2.0},
   url = {https://raw.githubusercontent.com/codemeta/codemeta/2.0/codemeta.jsonld},
   author = {Jones,  Matthew B and Boettiger,  Carl and Mayes,  Abby Cabunoc and {Arfon Smith} and Slaughter,  Peter and Niemeyer,  Kyle and Gil,  Yolanda and Fenner,  Martin and Nowak,  Krzysztof and Hahnel,  Mark and Coy,  Luke and Allen,  Alice and Crosas,  Mercè and Sands,  Ashley and Hong,  Neil Chue and Cruse,  Patricia and Katz,  Dan and Goble,  Carole},
-  title = {CodeMeta: an exchange schema for software metadata},
+  title = {{CodeMeta}: an exchange schema for software metadata},
   publisher = {KNB Data Repository},
   year = {2017}
 }
@@ -308,7 +308,7 @@ Code},
 }
 
 @Manual{xiao_liftr_2018,
-  title = {liftr: Containerize R Markdown Documents for Continuous Reproducibility},
+  title = {liftr: Containerize {R} {Markdown} Documents for Continuous Reproducibility},
   author = {Nan Xiao},
   year = {2018},
   note = {R package version 0.9},
@@ -316,7 +316,7 @@ Code},
 }
 
 @Manual{xie_xfun_2018,
[…]
