### 2019-08-21 #1603 @kyleniemeyer
Hi @nuest—I did some minor final copyediting, if you could merge these PRs: https://github.com/o2r-project/containerit/pull/166 and https://github.com/o2r-project/containerit/pull/167
