+++ paper.bib
@@ -56,8 +56,7 @@
 }
 
 @Manual{brokamp_automagic_2017,
-  title = {automagic: Automagically Document and Install Packages Necessary to Run {R}
-Code},
+  title = {automagic: Automagically Document and Install Packages Necessary to Run {R} Code},
   author = {Cole Brokamp},
   year = {2017},
   note = {R package version 0.3},
