### 2019-08-21 #1603 @nuest
Merged! Would you like me to update the Zenodo deposit (and get a new DOI) ?
