### 2019-08-21 #1603 @kyleniemeyer
@nuest that isn't necessary, since JOSS archives the final version of the paper itself
