### 2019-08-21 #1603 @kyleniemeyer
Congrats @nuest on your article's publication in JOSS! Many thanks to @terrytangyuan for editing, and @vsoch and @cole-brokamp for reviewing this submission.
